Patched a possible issue where  uses C++'s string 'length()' property but doesn't cast  to .  now casts the value to the  type

+++ src/core/program.cpp
@@ -883,7 +883,7 @@ void Program::resolve(const Instruction& ins) {
 
         Value out{};
         out.kind = ValueKind::Integer;
-        out.i = in.s.length();
+        out.i = static_cast<int64_t>(in.s.length());
 
         write_operand(x, out);
         pc++;
